backend/requirements.txt
passlib==1.7.2
pydantic==1.5.1
fastapi==0.58.1
PyJWT==1.7.1
python-multipart
sqlalchemy
flask_login
flask_script
flask
flask_admin
click
pyarrow==2.0.0
pandas
gunicorn
gevent
flask_sqlalchemy
flask_cors
redis
PyYaml
itsdangerous
werkzeug
